pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "twr-lm"
version = "0.1.0"
description = "PyTorch research scaffold for Tokenless Write-Think-Read Latent Machines"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "torch>=2.2",
  "torchvision>=0.17",
  "PyYAML>=6.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "src.twr*"]
